Rework dependency resolution
We also use pr-downloader rather than ad-hoc methods to fetch new versions
of the spring engine. Still using the old method for maps for now, but
pr-downloader supports fetching maps, too.

Most importantly, this should allow the process to automatically download
Zero-K versions.

This *might* make things parallel safe again, but you're going to need
a lot of resources to run multiple instances of springrts in parallel.

Note that this approach drastically increases the number of files that
need to be considered and read by the build process.