Factory progression rework
Track full factory progression, rather than just the first plop.

Still store winner_fac and loser_fac, for the sake of backwards
compatibility and simplicity for the matchups chart.

During data coercion, append a "dummy" entry for "Never" if one doesn't
already exist. This is presumed to always be terminal. This is necessary,
as dc assumes stem nodes cannot have any value of their own. We also
strip down to the first three factories placed, as anything more specific
is likely to be too low frequency to have any value, and would just add
more clutter to a chart that is already, frankly, pushing it a bit.

Expose the sunburst chart, similar to the pie chart.

Rework the sunburst chart implementation to use linear radial increments.

Move the mirror pie chart one column to the left.