Workaround for Spring RulesParam coercion
Zero-K uses {Get,Set}GameRulesParam for storing long and short versions
of allyteam names, which is used in reporting who won.

Unfortunately, Spring uses lua_isnumber when deciding what type to
store a parameter as, and lua_isnumber:
> Returns 1 if the value at the given acceptable index is a number or
> a string convertible to a number, and 0 otherwise.
Which means that even if ZK explicitly casts to a string beforehand, as
long as it looks like a number, Spring will helpfully store it as a
number.

This means that if a battle has a winner that is a player with a name
that looks like a very long number (long enough not to be stored
losslessly as a floating point), then we end up getting something
quite different from Spring.GetGameRulesParam.

For example, in https://zero-k.info/Battles/Detail/855246 we have
the winner as 1974402712, which anything trying to read back as a
RulesParam will see as 1974402688 due to the "helpful" yet lossy
coercion.

We work around this by, if fetching the winning player fails,
converting the keys in the name_to_player map through simulated
float truncation. This is pretty awful, but we can't use a fixed
engine version, or mod version that works around this, given a
replay for a specific mod and engine version.