Archive old stat data
This moves old demos, stats, and summaries folders for each replay to an
archive/ directory that does not need regular scanning. The contents of
summaries are stores in archive.json.frags so that the archive/
directory does not need regular scanning.

This means that the status script now only reports on progress from
non-archived data. No other public-facing changes should be visible.