Scrape detail.html for lobbyid as fallback
This is not reliable, as a player may have had their name changed between
when the battle occurred and when the detail.html was scraped.

This is not expected to be a common scenario, and postprocess.py will
raise a failing exit code when met. This allows for manual editing
of events.log to provide userid.

Place your bets on how many demofiles I'll need to zgrep now!