Horizontally partition data
All demos, stats, summaries are now stored under "shards". The shard ID is
the ID of the battle, divided by 10000.

Almost all places that an ID is currently used, the path SHARD/ID is now
used instead, with SHARD being a group a battle belongs to, and ID being
the full ID used.

Helper directories are created at shards/SHARD, which contain symlinks
to their relevant folder.

For example:

shards/160/demos -> ../../demos/160/
shards/160/stats -> ../../stats/160/
shards/160/summaries -> ../../summaries/160/

The structure within a shard is much the same as the previous top-level
structure, except the final product is now shard.json.frags instead of
all.json, and does not have its contents enclosed by array braces.

A top level all.json is still created, enclosing with array braces all
shard.json.frags files.

Each shard contains a "fetch-stamp". This is created by zkreplay-fetch when it
initially creates the folder, and is touched whenever a new replay is made
available. If a shard's timestamp is newer than its shard.json, or if
shard.json does not exist, a separate Makefile invocation is made to perform
all necessary processing in that shard. Most of the time, only the newest
shard or two will need processing, allowing the vast majority of files
to be skipped.

This transforms the process from being one where an appreciable fraction of a
million files need to be examined to being one where a few thousand need to be
examined.

This could still use a proper rewrite, but this fixes the scaling issues
without much expenditure of developer mana.

Various documentation and naming conventions have been cleaned up in the
process.

All references to the previous archiving system have been removed, as I
don't believe it was ever fully complete.